% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changomics.changomics_2.R
\name{changomics.2}
\alias{changomics.2}
\title{changomics.2}
\usage{
changomics.2(data)
}
\arguments{
\item{data}{input dataset of metabolites as data frame.}
}
\value{
the corrected dataframe of metabolite(s).
}
\description{
A function for metabolite correction. This function first estimate the change points
(jumps) and then correct each different segment by detrending and normalizing when needed.
}
\examples{
met1<-rnorm(1000)
changomics(as.data.frame(met1))
}
