% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changomics.changomics.R
\name{changomics}
\alias{changomics}
\title{changomics}
\usage{
changomics(data, max.knots = 10, debug = F, runall = F)
}
\arguments{
\item{data}{input dataset of metabolites as data frame.}

\item{max.knots}{Maximum number of knots as \% .}

\item{debug}{Debug mode. Defaults to TRUE}

\item{runall}{Apply correction regrdless of white noise test.Defaults to FALSE.}
}
\value{
the corrected dataframe of metabolite(s).
}
\description{
A function for metabolite correction. This function first estimate the change points
(jumps) and then correct each different segment by detrending and normalizing when needed.
}
\examples{
met1<-rnorm(200)
changomics(as.data.frame(met1))
}
