% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.normalize.winn.R
\name{correct.pp}
\alias{correct.pp}
\title{Apply WiNN correction to pooled plasma data and returns
CVs before and after correction.}
\usage{
correct.pp(winn.corrected, untrnsf.dt, pooled.plasma, stdy.id)
}
\arguments{
\item{winn.corrected}{The WiNN corrected metabolite dataset.}

\item{untrnsf.dt}{The uncorrected metabolite dataset (with \emph{NO} pooled plasma in it).}

\item{pooled.plasma}{The Pooled Plasma dataset}
}
\value{
list with CVs before and after correction
}
\description{
Applies WiNN correction to PP data and returns CVs
before and after correction.
}
\examples{
correct.pp()
}
\keyword{correct.pp}
