% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.normalize.winn.R
\name{cv.plot}
\alias{cv.plot}
\title{Creates before and after correction CV plot}
\usage{
cv.plot(cv.pp, cv.thr = NA, stdy.id)
}
\arguments{
\item{cv.pp}{The pooled plasma CV object returned by correct.pp}

\item{cv.thr}{The CV threshold value for the plot}
}
\description{
Creates before and after correction CV plot.
}
\examples{
cv.plot()
}
\keyword{cv.plot}
