% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winn.R
\name{winn}
\alias{winn}
\title{winn}
\usage{
winn(data, end.plates, graph = FALSE)
}
\arguments{
\item{data}{input dataset of metabolites as data frame.}

\item{end.plates}{vector containing the position of the changes in plates
in terms of run order. The vector must be ordered.}

\item{graph}{TRUE or FALSE. If TRUE will display the graph of uncorrected
data with plates and correction and the corrected signal on a
second plot.}
}
\value{
the corrected dataframe of metabolite(s).
}
\description{
A function for metabolite correction.
This function corrects batch effects and drifts
with the plate information by detrending and normalizing when needed.
}
\examples{
met1<-rnorm(200)
winn(as.data.frame(met1),end.plates=100)
}
