% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changomics.changomics_2.R
\name{changomics.2}
\alias{changomics.2}
\title{changomics.2}
\usage{
changomics.2(data, max.knots = 10, debug = T, runall = F)
}
\arguments{
\item{data}{input dataset of metabolites as data frame.}

\item{max.knots}{Maximum number of knots as \% .}

\item{debug}{Debug mode. Defaults to TRUE}

\item{runall}{Apply correction regrdless of white noise test.Defaults to FALSE.}
}
\value{
the corrected dataframe of metabolite(s).
}
\description{
A function for metabolite correction. This function first estimate the change points
(jumps) and then correct each different segment by detrending and normalizing when needed.
}
\examples{
met1<-rnorm(1000)
changomics(as.data.frame(met1))
}
