% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.df.R
\docType{data}
\name{met.df}
\alias{met.df}
\title{met.df}
\format{
an object of class data.frame with 2 simluated metabolites data

\describe{ In this dataset we are giving two metabolites data that we generated ourselves. "met1" is
just some white noise from a normal distribution and "met2" is a metaboite with jumps and trends to correct}
}
\usage{
data(met.df)
}
\description{
met.df
}
\examples{
data(met.df)
met.df<-as.data.frame(met.df)
corrected<-changomics.2(met.df)


}
\keyword{dataset}
