% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changomics.Winn.R
\name{winn}
\alias{winn}
\title{WiNN in development}
\usage{
winn(input.dat, group.var = "plate", max.knots = 10, debug = T, runall = F)
}
\arguments{
\item{input.dat}{input dataset as data frame or matrix}

\item{group.var}{Grouping variable. Defaults to "plate".}

\item{max.knots}{Maximum number of knots as \% .}

\item{debug}{Debug mode. Defaults to TRUE.}

\item{runall}{Apply correction regrdless of white noise test.Defaults to FALSE.}
}
\value{
A list (corrected.metabolites, correction.summary)
}
\description{
A function for metabolite correction
}
\keyword{WiNN}
