% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.normalize.winn.R
\name{winn}
\alias{winn}
\title{WiNN in development}
\usage{
winn(
  input.dat,
  stdy.id,
  group.var = "plate",
  max.knots = 10,
  debug = T,
  runall = F,
  save.corrected.data = T,
  save.uncorrected.data = T,
  save.pdfs = T
)
}
\arguments{
\item{input.dat}{Study ID.}

\item{group.var}{Grouping variable. Defaults to "plate".}

\item{max.knots}{Maximum number of knots as \% .}

\item{debug}{Debug mode. Defaults to TRUE.}

\item{runall}{Apply correction regrdless of white noise test.Defaults to FALSE.}

\item{save.corrected.data}{Save corrected dataset. Defaults to TRUE}

\item{save.pdfs}{Save pdfs of corrected/uncorrected data. Defaults to TRUE}
}
\value{
A list (corrected.strict, corrected.lenient,corrected.resid, correction.summary)
}
\description{
A function for metabolite correction
}
\examples{
winn()
}
\keyword{WiNN}
